removed plugin update to wicket 6.7.0 added some wicketstuff added some health stuff

+++ src/main/java/org/bricket/b4/health/entity/HealthSettings.java
@@ -45,7 +45,7 @@ public class HealthSettings extends DomainObject {
     private Long uid;
 
     @Column(name = "height")
-    private long height;
+    private Double height;
 
     @Column(name = "birthday")
     private Date birthday;
